refactor: Move flake8_tidy_imports Settings to mod.rs
